feat(core): introduce `PluginManager` and centralize plugin handling
- Added a new `PluginManager` class to encapsulate and streamline plugin management.
- Refactored existing code to replace `PluginUtils` with `PluginManager` in key areas.
- Updated CLI commands and tests to integrate with the new `PluginManager`.
- Improved plugin loading logic by supporting development mode manifests and enhanced class handling.
