+++ spellchecker/morfologik/src/test/java/org/omegat/spellchecker/morfologik/MorfologikSpellcheckerTest.java
@@ -31,7 +31,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Collections;
 
 import org.apache.commons.io.FileUtils;
 import org.junit.AfterClass;
@@ -42,7 +41,7 @@ import org.omegat.core.Core;
 import org.omegat.core.data.NotLoadedProject;
 import org.omegat.core.data.ProjectProperties;
 import org.omegat.core.spellchecker.ISpellChecker;
-import org.omegat.filters2.master.PluginUtils;
+import org.omegat.plugin.PluginManager;
 import org.omegat.util.Language;
 import org.omegat.util.TestPreferencesInitializer;
 
@@ -55,7 +54,7 @@ public class MorfologikSpellcheckerTest {
 
     @BeforeClass
     public static void setUpClass() throws Exception {
-        PluginUtils.loadPlugins(Collections.emptyMap());
+        PluginManager.loadPlugins();
         tmpDir = Files.createTempDirectory("omegat");
         assertThat(tmpDir.toFile()).isDirectory();
         configDir = Files.createDirectory(tmpDir.resolve(".omegat"));
